requirements.txt
requests
beautifulsoup4
paho-mqtt
PyYAML
flask